main badge added

+++ src/main/java/com/sorted/portal/assisting/beans/config/ProductBean.java
@@ -3,12 +3,14 @@ package com.sorted.portal.assisting.beans.config;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProductBean {
     private String name;
     private String id;
-    private double mrp;
-    private double sellingPrice;
+    private BigDecimal mrp;
+    private BigDecimal sellingPrice;
     private String image;
 }
